[NDGL-16] chore: build-logic 구조 수정
