+++ build-logic/src/main/kotlin/NDGLDataPlugin.kt
@@ -1,7 +1,6 @@
 import convention.configureCoroutineAndroid
 import convention.configureHiltAndroid
 import convention.configureKotlinAndroid
-import extensions.configureAndroidLibrary
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 
@@ -12,7 +11,6 @@ class NDGLDataPlugin : Plugin<Project> {
         }
 
         configureKotlinAndroid()
-        configureAndroidLibrary()
         configureHiltAndroid()
         configureCoroutineAndroid()
     }
