+++ build-logic/src/main/kotlin/NDGLFeaturePlugin.kt
@@ -2,7 +2,6 @@ import convention.configureComposeAndroid
 import convention.configureCoroutineAndroid
 import convention.configureFirebase
 import convention.configureHiltAndroid
-import extensions.configureAndroidLibrary
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.dependencies
@@ -11,14 +10,10 @@ import util.libs
 class NDGLFeaturePlugin : Plugin<Project> {
     override fun apply(target: Project): Unit = with(target) {
         with(pluginManager) {
-            apply("com.android.library")
-            apply("org.jetbrains.kotlin.android")
+            apply("ndgl.android.library")
         }
 
-        configureAndroidLibrary()
         configureHiltAndroid()
-        configureFirebase()
-        configureComposeAndroid()
         configureCoroutineAndroid()
 
         dependencies {
