+++ build-logic/src/main/kotlin/convention/KotlinAndroid.kt
@@ -20,6 +20,15 @@ internal fun Project.configureKotlinAndroid() {
         defaultConfig {
             minSdk = Configuration.MIN_SDK
         }
+        buildTypes {
+            getByName("release") {
+                isMinifyEnabled = false
+                proguardFiles(
+                    getDefaultProguardFile("proguard-android-optimize.txt"),
+                    "proguard-rules.pro",
+                )
+            }
+        }
         compileOptions {
             sourceCompatibility = Configuration.JAVA_VERSION
             targetCompatibility = Configuration.JAVA_VERSION
@@ -39,6 +48,7 @@ internal fun Project.configureKotlinAndroid() {
 
     dependencies {
         "coreLibraryDesugaring"(libs.findLibrary("android.desugarJdkLibs").get())
+        "detektPlugins"(libs.findLibrary("detekt.formatting").get())
     }
 }
 
@@ -51,7 +61,7 @@ internal fun Project.configureKotlin() {
             freeCompilerArgs.set(
                 freeCompilerArgs.get() + listOf(
                     "-opt-in=kotlin.RequiresOptIn",
-                )
+                ),
             )
         }
     }
